+++ packages/_vue3-migration-test/src/components/index.ts
@@ -17,4 +17,6 @@ export { default as TestBaseIdModal } from './modals/test-base-id-modal.vue';
 export { default as TestExtraParams } from './extra-params/test-extra-params.vue';
 export { default as TestSearch } from './test-search.vue';
 export { default as TestTagging } from './tagging/test-tagging.vue';
+export { default as TestRenderlessExtraParam } from './extra-params/test-renderless-extra-param.vue';
+export { default as TestIcons } from './icons/test-icons.vue';
 export { default as TestDisplayEmitter } from './test-display-emitter.vue';