+++ packages/_vue3-migration-test/src/x-modules/experience-controls/components/index.ts
@@ -1,0 +1,1 @@
+export { default as TestExperienceControls } from './test-experience-controls.vue';