+++ packages/x-components/src/components/icons/checkbox-unselected.vue
@@ -1,10 +1,5 @@
-<template functional>
-  <svg
-    :class="['x-icon'].concat(data.staticClass, data.class)"
-    viewBox="0 0 8 8"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
+<template>
+  <svg class="x-icon" viewBox="0 0 8 8" fill="none" xmlns="http://www.w3.org/2000/svg">
     <rect
       x="1"
       y="1"
@@ -17,7 +12,3 @@
     />
   </svg>
 </template>
-
-<script lang="ts">
-  export default {};
-</script>