+++ packages/x-components/src/components/icons/chevron-left.vue
@@ -1,10 +1,5 @@
-<template functional>
-  <svg
-    :class="['x-icon'].concat(data.staticClass, data.class)"
-    viewBox="0 0 8 8"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
+<template>
+  <svg class="x-icon" viewBox="0 0 8 8" fill="none" xmlns="http://www.w3.org/2000/svg">
     <path
       d="M5.19995 6.5L2.69995 4L5.19995 1.5"
       stroke="currentColor"
@@ -14,7 +9,3 @@
     />
   </svg>
 </template>
-
-<script lang="ts">
-  export default {};
-</script>