+++ packages/x-components/src/x-modules/extra-params/components/__tests__/renderless-extra-params.spec.ts
@@ -82,7 +82,7 @@ describe('testing RenderlessExtraParam component', () => {
     expect(userChangedExtraParamsCallback).toHaveBeenCalledWith<[WirePayload<Dictionary<unknown>>]>(
       {
         eventPayload: { warehouse: 45678 },
-        metadata: { moduleName: 'extraParams', location: undefined, replaceable: true }
+        metadata: { moduleName: 'extraParams', location: 'none', replaceable: true }
       }
     );